REFRACT: - Reorganizando componentes
