+++ frontend/src/components/common/ProfileEmployee.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="bg-[#dbeafe] min-h-screen">
         <EmployeeHeader v-if="!isEmployer" :user="user" />
-        <MainEmployerHeader v-else :user="user" />
+        <MainEmployerHeader v-else />
         
         <div class="mx-[171px] my-[41px] space-y-[41px] pb-[41px]">
             
@@ -258,10 +258,9 @@
 </template>
 
 <script>
-import "../assets/Neumorfismo.css";
-import EmployeeHeader from './common/EmployeeHeader.vue'
-import MainEmployerHeader from './common/MainEmployerHeader.vue'
-import apiConfig from '../config/api.js'   // <-- agregado
+import EmployeeHeader from '../common/EmployeeHeader.vue';
+import MainEmployerHeader from '../common/MainEmployerHeader.vue';
+import apiConfig from '../../config/api.js';
 
 export default {
   name: 'EditInfoEmployee',
@@ -376,8 +375,6 @@ export default {
           this.saving = false;
           return;
         }
-
-        // Validación de campos laborales si es empleador
         if (this.isEmployer && (
           !this.editedUserData.departamento.trim() ||
           !this.editedUserData.puesto.trim() ||
@@ -387,7 +384,6 @@ export default {
           this.saving = false;
           return;
         }
-
         const updateData = {
           nombre: this.editedUserData.nombre.trim(),
           segundoNombre: this.editedUserData.segundoNombre.trim(),
@@ -398,15 +394,10 @@ export default {
           distrito: this.editedUserData.distrito.trim(),
           direccionParticular: this.editedUserData.direccionParticular.trim(),
           iban: this.editedUserData.iban.trim(),
-
           departamento: this.editedUserData.departamento.trim(),
           puesto: this.editedUserData.puesto.trim(),
           salario: this.editedUserData.salario
-
         };
-
-        console.log('Enviando datos al backend:', updateData);
-
         const response = await fetch(apiConfig.endpoints.profileEmployee(this.id), {
           method: 'PUT',
           headers: {
@@ -414,18 +405,14 @@ export default {
           },
           body: JSON.stringify(updateData)
         });
-
         const result = await response.json();
-
         if (response.ok && result.success) {
           this.successMessage = result.message;
           this.isEditing = false;
-          
           await this.fetchUserData();
         } else {
           this.error = result.message || 'Error al guardar los cambios';
         }
-
       } catch (error) {
         console.error('Error al guardar cambios:', error);
         this.error = 'Error de conexión al servidor';
@@ -498,7 +485,6 @@ export default {
         }
       }
     },
-
 
     checkUserRole() {
       const userRaw = localStorage.getItem('user');
